chore(SettingsPage/Webhooks): move formatValue to the component using it
