+++ packages/core/admin/admin/src/pages/SettingsPage/pages/Webhooks/EditView/components/EventInput/tests/formatValue.test.js
@@ -1,4 +1,4 @@
-import formatValue from '../utils/formatValue';
+import { formatValue } from '..';
 
 describe('utils | formatValue', () => {
   it('should format array to object', () => {
