--- packages/core/admin/admin/src/pages/SettingsPage/pages/Webhooks/EditView/components/EventInput/utils/formatValue.js
@@ -1,13 +1,0 @@
-const formatValue = (value) =>
-  value.reduce((acc, curr) => {
-    const key = curr.split('.')[0];
-
-    if (!acc[key]) {
-      acc[key] = [];
-    }
-    acc[key].push(curr);
-
-    return acc;
-  }, {});
-
-export default formatValue;
